Adds rep to demystify prelude functions
I think I'd like to add more functions here in time, be there are all
the ones we have right now from Eden's reps. I could imagine that maybe
we might end up having DemystifyingPrelude be a re-occuring theme in the
reps as we tackle more advanced topics as well.