Organizes the reps into levels
Hopefully this will aid newcomes to the reps in getting started. I
purposely gave only one digit for ordering on these as a constraint to
force is to have no more than 10 reps per level (if we decide to use 0)
to keep solid sense of progress for anyone doing the reps from start to
end. Once we fill out the reps to Level9, we'll have to decide whether
it makes sense to break this constraint for levels or introduce a third
layer of organization to get us beyond the level ceiling.