Adds Set7_TourOfList_Part1 - a initial tour of the list api
Clearly a complete tour of the list api is going to need to be broken
across multiple sets of reps. This is simple a first cut at an initial
step on the tour that I formed by simply going through the list api
docs. Once we have more sections of the tour, we will probably find
items to move around between tour sections to make the experience more
cohesive.