Adds a first set of reps about typeclasses
I injected in front the sets in Level 2 that are typeclass heavy to make
sure the basic syntax is firm before getting to those reps.

There's also a couple of reps in Level 1 that define instances for Eq
and Ord, but I think I'm content to leave those where they are and let
them act as a preview of what is to come.